# proyecto/schema/openapi_liga.yaml
openapi: 3.0.0
info:
  title: API de Predicción de LaLiga
  description: API para gestionar y predecir los resultados de los partidos de LaLiga.
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Servidor local

components:
  schemas:
    Usuario:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
        password:
          type: string
        links:
          type: array
          items:
            $ref: '#/components/schemas/Link'
    Partido:
      type: object
      required:
        - matchId
        - teams
        - goals
      properties:
        id:
          type: string
        date:
          type: string
          format: date-time
        teams:
          type: object
          required:
            - home
            - away
          properties:
            home:
              type: object
              required:
                - id
                - name
                - winner
              properties:
                id:
                  type: integer
                name:
                  type: string
                logo:
                  type: string
                winner:
                  type: boolean
            away:
              type: object
              required:
                - id
                - name
                - winner
              properties:
                id:
                  type: integer
                name:
                  type: string
                logo:
                  type: string
                winner:
                  type: boolean
        goals:
          type: object
          required:
            - home
            - away
          properties:
            home:
              type: integer
            away:
              type: integer
        statistics:
          type: array
          items:
            type: object
            properties:
              team:
                type: object
                properties:
                  id:
                    type: integer
                  name:
                    type: string
                  logo:
                    type: string
              stats:
                type: object
                properties:
                  shotsOnGoal:
                    type: integer
                  shotsOffGoal:
                    type: integer
                  totalShots:
                    type: integer
                  blockedShots:
                    type: integer
                  shotsInsideBox:
                    type: integer
                  shotsOutsideBox:
                    type: integer
                  fouls:
                    type: integer
                  cornerKicks:
                    type: integer
                  offsides:
                    type: integer
                  possession:
                    type: string
                  yellowCards:
                    type: integer
                  redCards:
                    type: integer
                  goalkeeperSaves:
                    type: integer
                  totalPasses:
                    type: integer
                  accuratePasses:
                    type: integer
                  passPercentage:
                    type: string
                  expectedGoals:
                    type: string
        links:
          type: array
          items:
            $ref: '#/components/schemas/Link'
    Prediccion:
      type: object
      properties:
        equipoCasa:
          type: string
        equipoVisitante:
          type: string
        resultadoPrediccion:
          type: string
        links:
          type: array
          items:
            $ref: '#/components/schemas/Link'
    Equipo:
      type: object
      required:
        - teamId
        - name
        - venue
      properties:
        id:
          type: string
        teamId:
          type: integer
        name:
          type: string
        code:
          type: string
        country:
          type: string
        founded:
          type: integer
        logo:
          type: string
        venue:
          type: object
          required:
            - name
          properties:
            id:
              type: integer
            name:
              type: string
            address:
              type: string
            city:
              type: string
            capacity:
              type: integer
            image:
              type: string
        jugadores:
          type: array
          items:
            type: object
            properties:
              playerId:
                type: string
              name:
                type: string
              position:
                type: string
        links:
          type: array
          items:
            $ref: '#/components/schemas/Link'

    Jugador:
      type: object
      properties:
        id:
          type: string
        nombre:
          type: string
        equipoId:
          type: string
        equipoNombre:
          type: string
        posicion:
          type: string
        links:
          type: array
          items:
            $ref: '#/components/schemas/Link'
    Link:
      type: object
      properties:
        rel:
          type: string
        href:
          type: string
        method:
          type: string

paths:
  /usuarios/registro:
    post:
      summary: Registrar un nuevo usuario
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Usuario'
      responses:
        '201':
          description: Usuario registrado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuario'
        '400':
          description: Solicitud incorrecta
          content:
            application/json:
              example:
                message: "Datos de usuario no válidos"
        '500':
          description: Error interno del servidor
          content:
            application/json:
              example:
                message: "Error al registrar el usuario"
  /usuarios/login:
    post:
      summary: Iniciar sesión de usuario
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Usuario autenticado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuario'
        '400':
          description: Solicitud incorrecta
          content:
            application/json:
              example:
                message: "Credenciales no válidas"
        '401':
          description: No autorizado
          content:
            application/json:
              example:
                message: "Correo electrónico o contraseña incorrectos"
        '500':
          description: Error interno del servidor
          content:
            application/json:
              example:
                message: "Error al iniciar sesión del usuario"
  /partidos:
    get:
      summary: Obtener todos los partidos
      responses:
        '200':
          description: Lista de partidos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Partido'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              example:
                message: "Error al obtener los partidos"
    post:
      summary: Añadir un nuevo partido
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Partido'
      responses:
        '201':
          description: Partido creado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Partido'
        '400':
          description: Solicitud incorrecta
          content:
            application/json:
              example:
                message: "Datos del partido no válidos"
        '500':
          description: Error interno del servidor
          content:
            application/json:
              example:
                message: "Error al crear el partido"
  /partidos/{id}:
    get:
      summary: Obtener un partido específico por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Detalles del partido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Partido'
        '404':
          description: Partido no encontrado
          content:
            application/json:
              example:
                message: "Partido no encontrado"
        '500':
          description: Error interno del servidor
          content:
            application/json:
              example:
                message: "Error al obtener el partido"
    put:
      summary: Actualizar un partido
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Partido'
      responses:
        '200':
          description: Partido actualizado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Partido'
        '400':
          description: Solicitud incorrecta
          content:
            application/json:
              example:
                message: "Datos del partido no válidos"
        '404':
          description: Partido no encontrado
          content:
            application/json:
              example:
                message: "Partido no encontrado"
        '500':
          description: Error interno del servidor
          content:
            application/json:
              example:
                message: "Error al actualizar el partido"
    delete:
      summary: Eliminar un partido
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Partido eliminado exitosamente
        '404':
          description: Partido no encontrado
          content:
            application/json:
              example:
                message: "Partido no encontrado"
        '500':
          description: Error interno del servidor
          content:
            application/json:
              example:
                message: "Error al eliminar el partido"
  /partidos/{id}/estadisticas:
    get:
      summary: Obtener estadísticas de un partido específico por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Estadísticas del partido
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  estadisticas:
                    type: object
        '404':
          description: Partido no encontrado
          content:
            application/json:
              example:
                message: "Partido no encontrado"
        '500':
          description: Error interno del servidor
          content:
            application/json:
              example:
                message: "Error al obtener las estadísticas del partido"
  /predicciones:
    post:
      summary: Obtener predicción para un partido
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                equipoCasa:
                  type: string
                equipoVisitante:
                  type: string
      responses:
        '200':
          description: Resultado de la predicción
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Prediccion'
        '400':
          description: Solicitud incorrecta
          content:
            application/json:
              example:
                message: "Datos de equipos no válidos"
        '500':
          description: Error interno del servidor
          content:
            application/json:
              example:
                message: "Error al obtener la predicción"
  /equipos:
    get:
      summary: Obtener todos los equipos
      responses:
        '200':
          description: Lista de equipos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Equipo'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              example:
                message: "Error al obtener los equipos"
    post:
      summary: Añadir un nuevo equipo
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Equipo'
      responses:
        '201':
          description: Equipo creado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Equipo'
        '400':
          description: Solicitud incorrecta
          content:
            application/json:
              example:
                message: "Datos del equipo no válidos"
        '500':
          description: Error interno del servidor
          content:
            application/json:
              example:
                message: "Error al crear el equipo"
  /equipos/{id}:
    get:
      summary: Obtener un equipo específico por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Detalles del equipo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Equipo'
        '404':
          description: Equipo no encontrado
          content:
            application/json:
              example:
                message: "Equipo no encontrado"
        '500':
          description: Error interno del servidor
          content:
            application/json:
              example:
                message: "Error al obtener el equipo"
    put:
      summary: Actualizar un equipo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Equipo'
      responses:
        '200':
          description: Equipo actualizado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Equipo'
        '400':
          description: Solicitud incorrecta
          content:
            application/json:
              example:
                message: "Datos del equipo no válidos"
        '404':
          description: Equipo no encontrado
          content:
            application/json:
              example:
                message: "Equipo no encontrado"
        '500':
          description: Error interno del servidor
          content:
            application/json:
              example:
                message: "Error al actualizar el equipo"
    delete:
      summary: Eliminar un equipo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Equipo eliminado exitosamente
        '404':
          description: Equipo no encontrado
          content:
            application/json:
              example:
                message: "Equipo no encontrado"
        '500':
          description: Error interno del servidor
          content:
            application/json:
              example:
                message: "Error al eliminar el equipo"
  /equipos/{id}/jugadores:
    get:
      summary: Obtener todos los jugadores de un equipo específico
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Lista de jugadores del equipo
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Jugador'
        '404':
          description: Equipo no encontrado
          content:
            application/json:
              example:
                message: "Equipo no encontrado"
        '500':
          description: Error interno del servidor
          content:
            application/json:
              example:
                message: "Error al obtener los jugadores del equipo"
  /jugadores:
    get:
      summary: Obtener todos los jugadores
      responses:
        '200':
          description: Lista de jugadores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Jugador'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              example:
                message: "Error al obtener los jugadores"
    post:
      summary: Añadir un nuevo jugador
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Jugador'
      responses:
        '201':
          description: Jugador creado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Jugador'
        '400':
          description: Solicitud incorrecta
          content:
            application/json:
              example:
                message: "Datos del jugador no válidos"
        '500':
          description: Error interno del servidor
          content:
            application/json:
              example:
                message: "Error al crear el jugador"
  /jugadores/{id}:
    get:
      summary: Obtener un jugador específico por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Detalles del jugador
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Jugador'
        '404':
          description: Jugador no encontrado
          content:
            application/json:
              example:
                message: "Jugador no encontrado"
        '500':
          description: Error interno del servidor
          content:
            application/json:
              example:
                message: "Error al obtener el jugador"
    put:
      summary: Actualizar un jugador
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Jugador'
      responses:
        '200':
          description: Jugador actualizado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Jugador'
        '400':
          description: Solicitud incorrecta
          content:
            application/json:
              example:
                message: "Datos del jugador no válidos"
        '404':
          description: Jugador no encontrado
          content:
            application/json:
              example:
                message: "Jugador no encontrado"
        '500':
          description: Error interno del servidor
          content:
            application/json:
              example:
                message: "Error al actualizar el jugador"
    delete:
      summary: Eliminar un jugador
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Jugador eliminado exitosamente
        '404':
          description: Jugador no encontrado
          content:
            application/json:
              example:
                message: "Jugador no encontrado"
        '500':
          description: Error interno del servidor
          content:
            application/json:
              example:
                message: "Error al eliminar el jugador"